requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bibtexparser==2.0.0b7
braceexpand==0.1.7
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clldutils==3.22.1
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.0
csvw==3.3.0
cycler==0.12.1
Cython==3.0.8
datasets==2.17.1
decorator==5.1.1
dill==0.3.8
dlinfo==1.2.1
docopt==0.6.2
editdistance==0.8.1
einops==0.7.0
executing==2.0.1
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
greenlet==3.0.3
grpcio==1.62.0
huggingface-hub==0.20.3
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.6
inflect==7.0.0
install==1.3.5
ipython==8.22.1
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jiwer==3.0.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.5
language-tags==1.2.0
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.0.post0
lightning-utilities==0.10.1
llvmlite==0.42.0
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
mutagen==1.47.0
nemo_toolkit==1.22.0
networkx==3.2.1
numba==0.59.0
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.15.0
optuna==3.5.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
phonemizer==3.2.1
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
pylatexenc==2.10
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
rdflib==7.0.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
segments==2.2.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.27
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.1
torchmetrics==1.3.1
torchvision==0.17.1
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
triton==2.2.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webdataset==0.2.86
websockets==12.0
Werkzeug==3.0.1
wget==3.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
youtokentome==1.0.6
yt-dlp==2023.12.30
